requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-SocketIO==5.5.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
netifaces==0.10.6
pyserial==3.5
python-engineio==4.11.2
python-socketio==5.12.1
setuptools==3.3
simple-websocket==1.1.0
socketio==0.2.1
Werkzeug==3.1.3
wsproto==1.2.0
